conditionnally show example picker
