+++ components/editor/modules/chart/index.js
@@ -65,8 +65,8 @@ export default ({ rule, subModules, TYPE }) => {
             m.TYPE === CANVAS_TYPE
               ? {
                   isNew: true,
-                  config: { type: 'Bar', y: 'label' },
-                  values: `label,value\nA,${random()}\nB,${random()}`
+                  config: {},
+                  values: ''
                 }
               : undefined
         })
